enhance: add functionality to only checkout a subdirectory of the repo
This is useful for module references like: `github.com/terraform-aws-modules/terraform-aws-rds//modules/db_instance`,
and means the whole repository doesn't need to be pulled if just a subdirectory is needed.

This doesn't handle any references within that submodule. For example, if it references modules located in other directories in the repo or if it symlinks into any other directories in the repo.